Refactor ElectricityChart component and update styling of EstimateOptions component
