+++ client/src/components/ElectricityChart/chartData.js
@@ -1,0 +1,59 @@
+const labels = [
+  '🇺🇸 united states',
+  '🇦🇺 australia',
+  '🇨🇦 canada',
+  '🇩🇪 germany',
+  '🇬🇧 united kingdom',
+  '🇷🇺 russia',
+  '🇪🇸 spain',
+  '🇪🇺 eu',
+  '🇿🇦 south africa',
+  '🇮🇹 italy',
+  '🌎 world',
+  '🇫🇷 france',
+  '🇯🇵 japan',
+  '🇧🇷 brazil',
+  '🇲🇽 mexico',
+  '🇨🇳 china',
+  '🇮🇳 india',
+  '🇳🇬 nigeria',
+  '',
+  '🏠 your home',
+];
+
+export const chartData = {
+  labels,
+  datasets: [
+    {
+      label: 'yearly co2 emissions',
+      data: [
+        2918, 2176, 1544, 1056, 971, 851, 784, 732, 719, 589, 569, 348, 303,
+        300, 296, 221, 147, 93, 0,
+      ],
+      backgroundColor: [
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 180, 0)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(0, 255, 0)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(255, 99, 132, 0.25)',
+        'rgba(0, 255, 255)',
+      ],
+      borderColor: ['rgba(0,0,0,.5)'],
+      borderWidth: 0.6,
+    },
+  ],
+};
