+++ client/src/components/EstimateOptions/EstimateOptions.scss
@@ -1,0 +1,60 @@
+@use '../../styles/partials/colors' as theme;
+@use '../../styles/partials/fonts' as f;
+@use '../../styles/partials/breakpoints' as bp;
+@use '../../styles/partials/mixins' as *;
+
+.aside {
+  width: 15%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+
+  &__link {
+    text-decoration: none;
+    height: 100%;
+    background-color: rgba(255, 255, 255, 0.5);
+
+    &.active {
+      background-color: theme.$gw-light-grey;
+
+      & .aside__card {
+        border-right: 0;
+        border-left: 10px solid theme.$gw-green;
+      }
+
+      & .aside__card-title {
+        color: theme.$gw-black;
+        font-weight: 600;
+      }
+    }
+  }
+
+  &__card {
+    height: 100%;
+
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border: 1px solid theme.$gw-white;
+    border-right: -3px;
+    border-left: 0;
+    // box-sizing: content-box;
+
+    &-title {
+      @include f.labels-links-buttons('Titillium Web');
+      color: theme.$gw-slate;
+      // font-weight: 400;
+
+      &:hover {
+        color: theme.$gw-black;
+      }
+    }
+
+    // Changes styling of Warehouse or Inventory buttons when active, depending on url
+  }
+
+  &__link {
+    cursor: pointer;
+  }
+}
