+++ client/src/components/Layout/Layout.scss
@@ -3,21 +3,15 @@
 @use '../../styles/partials/breakpoints' as bp;
 @use '../../styles/partials/mixins' as *;
 
-.layout {
+.main {
   display: flex;
   justify-content: center;
   flex-direction: column;
-
-  &__box {
-    width: 100vw;
-    // box-shadow: 0 0.2rem 0.5rem rgba(0, 0, 0, 0.15);
-    // border-radius: 0.4rem;
-  }
 }
 
 @include bp.tablet {
-  .layout {
-    &__box {
+  .main {
+    &__layout-box {
       width: 100%;
     }
   }
