+++ client/src/pages/estimate/Flights.js
@@ -1,12 +1,17 @@
 import React, { Component } from 'react';
 import Layout from '../../components/Layout/Layout';
-import EstimateTable from '../../components/EstimateTable/EstimateTable';
+import EstimateOptions from '../../components/EstimateOptions/EstimateOptions';
 
 export default class Flights extends Component {
   render() {
     return (
       <Layout>
-        <EstimateTable />
+        <section className="estimate__main-box">
+          <EstimateOptions />
+          <section className="estimate__content-box">
+            <div className="estimate__content"></div>
+          </section>
+        </section>
       </Layout>
     );
   }
