+++ client/src/pages/estimate/index.js
@@ -1,12 +1,17 @@
 import React, { Component } from 'react';
-import EstimateTable from '../../components/EstimateTable/EstimateTable';
+import EstimateOptions from '../../components/EstimateOptions/EstimateOptions';
 import Layout from '../../components/Layout/Layout';
 
 export default class Estimate extends Component {
   render() {
     return (
       <Layout>
-        <EstimateTable />
+        <section className="estimate__main-box">
+          <EstimateOptions />
+          <section className="estimate__content-box">
+            <div className="estimate__content"></div>
+          </section>
+        </section>
       </Layout>
     );
   }
